Reorg, add multiple_choice mod to voting package
