+++ packages/voting/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod deposit;
 pub mod error;
+pub mod multiple_choice;
 pub mod pre_propose;
 pub mod proposal;
 pub mod reply;
